WIP: Branch 4.5 again
This patch is not intended to be merged to master branch. Instead, it's
intended to be merged to a new 4.5 branch, replacing current ovirt-4.5,
which would be branched from current master.

It should represent all the intended differences between 4.5 and master.

The immediate reason is
https://github.com/oVirt/ovirt-appliance/issues/94 .

While working on cherry-picking the patches there to 4.5, I realized
there are many more patches in master not cherry-picked to 4.5, and at
least some of those I checked, made sense to also be in 4.5. So current
patch is an alternative to this pick-and-choose process.

Change-Id: Id6ba5fd06ef40e403a4e085aecfb64ec84b5056c
Signed-off-by: Yedidyah Bar David <didi@redhat.com>